@@ -1,9 +1,9 @@
 {
   "name": "nccwsc-projects",
-  "version": "0.0.1",
+  "version": "1.0.0",
   "license": "MIT",
   "engines": {
-    "node": ">=10.23.0"
+    "node": ">=18.20.4"
   },
   "scripts": {
     "ng": "ng",
@@ -15,54 +15,39 @@
   },
   "private": true,
   "dependencies": {
-    "@akveo/ng2-completer": "^9.0.1",
-    "@angular/animations": "^15.2.10",
-    "@angular/common": "^15.2.10",
-    "@angular/compiler": "^15.2.10",
-    "@angular/core": "^15.2.10",
-    "@angular/forms": "^15.2.10",
-    "@angular/localize": "^15.2.10",
-    "@angular/platform-browser": "^15.2.10",
-    "@angular/platform-browser-dynamic": "^15.2.10",
-    "@angular/router": "^15.2.10",
-    "@asymmetrik/ngx-leaflet": "3.1.0",
-    "@fortawesome/angular-fontawesome": "^0.12.0",
-    "@fortawesome/fontawesome-svg-core": "^1.2.36",
-    "@fortawesome/free-solid-svg-icons": "^5.15.4",
-    "@ng-bootstrap/ng-bootstrap": "^14.0.0",
-    "@ng-select/ng-select": "^10.0.0",
-    "@popperjs/core": "^2.10.2",
-    "ng2-smart-table": "^1.7.0",
-    "core-js": "^2.6.11",
-    "jquery": "^3.4.1",
-    "leaflet": "^1.7.1",
-    "ng2-completer": "^9.0.1",
-    "rxjs": "^6.6.3",
-    "rxjs-compat": "^6.6.3",
-    "tslib": "^2.0.0",
-    "umami": "^0.1.0",
-    "x2js": "3.1.1",
-    "zone.js": "~0.11.5"
+    "@angular/animations": "^18.1.0",
+    "@angular/common": "^18.1.0",
+    "@angular/compiler": "^18.1.0",
+    "@angular/core": "^18.1.0",
+    "@angular/forms": "^18.1.0",
+    "@angular/localize": "^18.1.0",
+    "@angular/material": "^18.1.0",
+    "@angular/platform-browser": "^18.1.0",
+    "@angular/platform-browser-dynamic": "^18.1.0",
+    "@angular/router": "^18.1.0",
+    "@asymmetrik/ngx-leaflet": "^18.0.1",
+    "@fortawesome/angular-fontawesome": "^0.15.0",
+    "@fortawesome/free-solid-svg-icons": "^6.5.2",
+    "@ng-bootstrap/ng-bootstrap": "^17.0.0",
+    "@ng-select/ng-select": "^13.4.1",
+    "@popperjs/core": "^2.11.8",
+    "core-js": "^3.37.1",
+    "jquery": "^3.7.1",
+    "leaflet": "^1.9.4",
+    "rxjs": "^7.8.1",
+    "x2js": "3.4.4",
+    "zone.js": "~0.14.7"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^15.2.11",
-    "@angular/cli": "^15.2.11",
-    "@angular/compiler-cli": "^15.2.10",
-    "@types/jasmine": "~3.6.0",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/leaflet": "1.5.19",
-    "@types/node": "^12.11.1",
-    "codelyzer": "^6.0.0",
-    "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "^6.3.19",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
-    "protractor": "~7.0.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~6.1.0",
-    "typescript": "~4.9.5"
+    "@angular-devkit/build-angular": "^18.1.0",
+    "@angular/cli": "^18.1.0",
+    "@angular/compiler-cli": "^18.1.0",
+    "@types/leaflet": "1.9.12",
+    "@types/node": "^20.14.10",
+    "angular-eslint": "18.1.0",
+    "eslint": "^9.6.0",
+    "ts-node": "~10.9.2",
+    "typescript": "~5.5.3",
+    "typescript-eslint": "8.0.0-alpha.38"
   }
 }
