@@ -1,4 +1,4 @@
-import "./polyfills.ts";
+import "./polyfills";
 
 import { platformBrowserDynamic } from "@angular/platform-browser-dynamic";
 import { enableProdMode } from "@angular/core";
