@@ -4,20 +4,20 @@
 import "@angular/localize/init";
 // This file includes polyfills needed by Angular and is loaded before
 // the app. You can add your own extra polyfills to this file.
-import "core-js/es6/symbol";
-import "core-js/es6/object";
-import "core-js/es6/function";
-import "core-js/es6/parse-int";
-import "core-js/es6/parse-float";
-import "core-js/es6/number";
-import "core-js/es6/math";
-import "core-js/es6/string";
-import "core-js/es6/date";
-import "core-js/es6/array";
-import "core-js/es6/regexp";
-import "core-js/es6/map";
-import "core-js/es6/set";
-import "core-js/es6/reflect";
+import "core-js/es/symbol";
+import "core-js/es/object";
+import "core-js/es/function";
+import "core-js/es/parse-int";
+import "core-js/es/parse-float";
+import "core-js/es/number";
+import "core-js/es/math";
+import "core-js/es/string";
+import "core-js/es/date";
+import "core-js/es/array";
+import "core-js/es/regexp";
+import "core-js/es/map";
+import "core-js/es/set";
+import "core-js/es/reflect";
+import "core-js/proposals/reflect-metadata";
 
-import "core-js/es7/reflect";
-import "zone.js/dist/zone";
+import "zone.js";
